backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-openai
langchain-community
langgraph
chromadb==0.4.24
pydantic>=2.10.0
pydantic-settings
boto3
tiktoken
sse-starlette
numpy
